# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Biblioteca",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "UsuariosJWT"
                ],
                "summary": "Añadir Usuario",
                "operationId": "5b1b8396e0f0688bc98f8183fce8e920",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "query",
                        "description": "Nombre de usuario nuevo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email de usuario nuevo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "contraseña",
                        "in": "query",
                        "description": "Contraseña de usuario nuevo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "contraseña_confirmation",
                        "in": "query",
                        "description": "Confirmacion de Contraseña de usuario nuevo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Añadir Usuario y te devuelve el usuario insertado."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "UsuariosJWT"
                ],
                "summary": "Logear Usuario",
                "operationId": "0e84dcf529e9de27079b2baa9301088c",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email de usuario ",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Contraseña de usuario",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Devuelve Access Token"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/auth/profile": {
            "get": {
                "tags": [
                    "UsuariosJWT"
                ],
                "summary": "Ver Perfil de Usuario por su token",
                "operationId": "0c6539310310b555855677fca7183f7f",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Introducir el token precedido de Bearer ",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Ver Perfil por token"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "UsuariosJWT"
                ],
                "summary": "Logout del Usuario",
                "operationId": "2956ff1a61d616bdb3be1e1c19747fd4",
                "responses": {
                    "201": {
                        "description": "Se sale de la sesion"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "UsuariosJWT"
                ],
                "summary": "Renovar token de Usuario",
                "operationId": "5675cfcbb73fb2f76b3f1acd6aeefed0",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Introducir el token precedido de Bearer ",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Ver token nuevo"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
...
    }
}